HHH-13154 Add support for JPA criteria on stateless sessions
