+++ hibernate-core/src/main/java/org/hibernate/query/criteria/internal/compile/CriteriaCompiler.java
@@ -14,11 +14,10 @@ import java.util.Map;
 import javax.persistence.TypedQuery;
 import javax.persistence.criteria.ParameterExpression;
 
-import org.hibernate.SessionFactory;
 import org.hibernate.dialect.Dialect;
 import org.hibernate.engine.jdbc.spi.JdbcServices;
 import org.hibernate.engine.spi.SessionFactoryImplementor;
-import org.hibernate.engine.spi.SessionImplementor;
+import org.hibernate.engine.spi.SharedSessionContractImplementor;
 import org.hibernate.internal.util.StringHelper;
 import org.hibernate.internal.util.collections.Stack;
 import org.hibernate.internal.util.collections.StandardStack;
@@ -38,9 +37,9 @@ import org.hibernate.type.Type;
  * @author Steve Ebersole
  */
 public class CriteriaCompiler implements Serializable {
-	private final SessionImplementor entityManager;
+	private final SharedSessionContractImplementor entityManager;
 
-	public CriteriaCompiler(SessionImplementor entityManager) {
+	public CriteriaCompiler(SharedSessionContractImplementor entityManager) {
 		this.entityManager = entityManager;
 	}
 
@@ -55,7 +54,7 @@ public class CriteriaCompiler implements Serializable {
 		final Map<ParameterExpression<?>, ExplicitParameterInfo<?>> explicitParameterInfoMap = new HashMap<>();
 		final List<ImplicitParameterBinding> implicitParameterBindings = new ArrayList<>();
 
-		final SessionFactoryImplementor sessionFactory = entityManager.getSessionFactory();
+		final SessionFactoryImplementor sessionFactory = entityManager.getFactory();
 
 		final LiteralHandlingMode criteriaLiteralHandlingMode = sessionFactory
 				.getSessionFactoryOptions()
